# DDIS_HR/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8000
  description: Generated server url
paths:
...
  /review/manager/performance/{performanceId}/evaluate:
    post:
      tags:
      - review-controller
      operationId: managerEvaluate
      parameters:
      - name: performanceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PerformanceReviewDTO'
...
  /api/password-reset/request:
    post:
      tags:
      - password-reset-controller
      operationId: request
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TokenResp'
...
  /review/{teamId}:
    get:
      tags:
      - review-controller
      operationId: getTeamReview
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeReviewDTO'
  /review/{reviewerId}/performance:
    get:
      tags:
      - review-controller
      operationId: getMyReviews
      parameters:
      - name: reviewerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewedPerformanceDTO'
  /review/history/{employeeId}:
    get:
      tags:
      - review-controller
      operationId: getHistoryByEmployeeId
      parameters:
      - name: employeeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewHistoryDTO'
...
components:
  schemas:
...
    EvaluationRequest:
      type: object
      properties:
        reviewerScore:
          type: integer
          format: int32
        reviewerContent:
          type: string
        employeeIdReviewer:
          type: integer
          format: int64
        decision:
          type: string
    PerformanceReviewDTO:
      type: object
      properties:
        performanceId:
          type: integer
          format: int64
        goalId:
          type: integer
          format: int64
        goalTitle:
          type: string
        performanceValue:
          type: number
          format: double
        reviewerScore:
          type: integer
          format: int32
        reviewerContent:
          type: string
        reviewerCreatedAt:
          type: string
          format: date-time
...
    Req:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        email:
          type: string
    TokenResp:
      type: object
      properties:
        tokenId:
          type: string
          format: uuid
...
    EmployeeReviewDTO:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        positionId:
          type: integer
          format: int64
        positionName:
          type: string
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        rankId:
          type: integer
          format: int64
        rankName:
          type: string
        goals:
          type: array
          items:
            $ref: '#/components/schemas/GoalDTO'
    GoalDTO:
      type: object
      properties:
        goalId:
          type: integer
          format: int64
        goalTitle:
          type: string
        goalContent:
          type: string
        goalValue:
          type: number
          format: double
        goalWeight:
          type: number
          format: double
        goalCreatedAt:
          type: string
          format: date
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        performance:
          $ref: '#/components/schemas/PerformanceDTO'
        performances:
          type: array
          items:
            $ref: '#/components/schemas/PerformanceDTO'
    PerformanceDTO:
      type: object
      properties:
        performanceId:
          type: integer
          format: int64
        selfreviewCreateAt:
          type: string
          format: date-time
        performanceValue:
          type: integer
          format: int32
        selfreviewContent:
          type: string
        attachmentKeys:
          type: array
          items:
            type: string
        attachmentFileNames:
          type: array
          items:
            type: string
        attachmentFileTypes:
          type: array
          items:
            type: string
        attachmentFileSizes:
          type: array
          items:
            type: integer
            format: int64
        reviewerScore:
          type: integer
          format: int32
        reviewerContent:
          type: string
        employeeIdReviewer:
          type: integer
          format: int64
    ReviewedPerformanceDTO:
      type: object
      properties:
        performanceId:
          type: integer
          format: int64
        goalId:
          type: integer
          format: int64
        goalTitle:
          type: string
        goalValue:
          type: number
          format: double
        performanceValue:
          type: number
          format: double
        reviewerScore:
          type: integer
          format: int32
        reviewerContent:
          type: string
        reviewerCreatedAt:
          type: string
          format: date-time
    ReviewHistoryDTO:
      type: object
      properties:
        yearMonth:
          type: string
        performanceDescription:
          type: string
        goalValue:
          type: integer
          format: int32
        performanceValue:
          type: integer
          format: int32
        selfReviewContent:
          type: string
        reviewScore:
          type: integer
          format: int32
...